shamu: introduce a much less crippled liblights
Each color (RGB) of the tri-color LED sits on it's own GPIO, which
is a rather unique design. As such, the leds-qpnp driver is not
currently designed for robust LED controls in this configuration.
Instead we convert to hard values of red, green, or blue, with no
blink support. Default notification color is green, low battery is
red, medium battery is blue, and battery full is green. Notification
LED takes precedence over battery LED, however when battery is
charging at full capacity both will be green.

This hal introduces basic user controls including custom charge and
notification colors. Blink and breathe are not supported at this time.

Change-Id: Ica9bc2aa820d1f7182e56aea54a66f7d1f9cde79

Conflicts:
	init.shamu.rc
	liblight/lights.c

Conflicts:
	init.shamu.rc
	liblight/lights.c
	overlay-cm/frameworks/base/core/res/res/values/config.xml